Adding error handling
We track more and more errors in sentry where `$childItem` is false and therefore the following code in `readAttribute()` is going to fail with `Call to a member function getProduct() on bool `. I didn't debug the whole stack trace yet. Nevertheless, the `current` function can return a bool and therefore the code afterwards won't succeed. 

Would appreciate a quick merge.